restructure + cleanup
