+++ src/QDL/saharaDefs.js
@@ -1,19 +1,19 @@
 import { structHelper_io } from "./utils"
 
 export const cmd_t = {
-  SAHARA_HELLO_REQ                : 0x1,
-  SAHARA_HELLO_RSP                : 0x2,
-  SAHARA_READ_DATA                : 0x3,
-  SAHARA_END_TRANSFER             : 0x4,
-  SAHARA_DONE_REQ                 : 0x5,
-  SAHARA_DONE_RSP                 : 0x6,
-  SAHARA_RESET_RSP                : 0x8,
-  SAHARA_CMD_READY                : 0xB,
-  SAHARA_SWITCH_MODE              : 0xC,
-  SAHARA_EXECUTE_REQ              : 0xD,
-  SAHARA_EXECUTE_RSP              : 0xE,
-  SAHARA_EXECUTE_DATA             : 0xF,
-  SAHARA_64BIT_MEMORY_READ_DATA   : 0x12,
+  SAHARA_HELLO_REQ : 0x1,
+  SAHARA_HELLO_RSP : 0x2,
+  SAHARA_READ_DATA : 0x3,
+  SAHARA_END_TRANSFER : 0x4,
+  SAHARA_DONE_REQ : 0x5,
+  SAHARA_DONE_RSP : 0x6,
+  SAHARA_RESET_RSP : 0x8,
+  SAHARA_CMD_READY : 0xB,
+  SAHARA_SWITCH_MODE : 0xC,
+  SAHARA_EXECUTE_REQ : 0xD,
+  SAHARA_EXECUTE_RSP : 0xE,
+  SAHARA_EXECUTE_DATA : 0xF,
+  SAHARA_64BIT_MEMORY_READ_DATA : 0x12,
 }
 
 export const exec_cmd_t = {
@@ -21,12 +21,12 @@ export const exec_cmd_t = {
 }
 
 export const sahara_mode_t = {
-  SAHARA_MODE_IMAGE_TX_PENDING  : 0x0,
-  SAHARA_MODE_COMMAND           : 0x3
+  SAHARA_MODE_IMAGE_TX_PENDING : 0x0,
+  SAHARA_MODE_COMMAND : 0x3
 }
 
 export const status_t = {
-  SAHARA_STATUS_SUCCESS  : 0x00,  // Invalid command received in current state
+  SAHARA_STATUS_SUCCESS : 0x00,  // Invalid command received in current state
   SAHARA_NAK_INVALID_CMD : 0x01,  // Protocol mismatch between host and targe
 }
 
@@ -45,18 +45,18 @@ export class CommandHandler {
       throw new Error("DataError!");
     let st = new structHelper_io(data);
     return {
-      cmd                     : st.dword(),
-      len                     : st.dword(),
-      version                 : st.dword(),
-      version_supported       : st.dword(),
-      cmd_packet_length       : st.dword(),
-      mode                    : st.dword(),
-      reserved1               : st.dword(),
-      reserved2               : st.dword(),
-      reserved3               : st.dword(),
-      reserved4               : st.dword(),
-      reserved5               : st.dword(),
-      reserved6               : st.dword(),
+      cmd : st.dword(),
+      len : st.dword(),
+      version : st.dword(),
+      version_supported : st.dword(),
+      cmd_packet_length : st.dword(),
+      mode : st.dword(),
+      reserved1 : st.dword(),
+      reserved2 : st.dword(),
+      reserved3 : st.dword(),
+      reserved4 : st.dword(),
+      reserved5 : st.dword(),
+      reserved6 : st.dword(),
     }
   }
 
@@ -66,9 +66,9 @@ export class CommandHandler {
       throw new Error("DataError!");
     let st = new structHelper_io(data);
     return {
-      cmd             : st.dword(),
-      len             : st.dword(),
-      image_id        : st.dword(),
+      cmd : st.dword(),
+      len : st.dword(),
+      image_id : st.dword(),
       image_tx_status : st.dword(),
     }
   }
@@ -79,8 +79,8 @@ export class CommandHandler {
       throw new Error("DataError");
     let st = new structHelper_io(data);
     return {
-      cmd             : st.dword(),
-      len             : st.dword(),
+      cmd : st.dword(),
+      len : st.dword(),
       image_tx_status : st.dword()
     }
   }
@@ -91,11 +91,11 @@ export class CommandHandler {
       throw new Error("DataError");
     let st = new structHelper_io(data)
     return {
-      cmd           : st.dword(),
-      len           : st.dword(),
-      image_id      : Number(st.qword()),
-      data_offset   : Number(st.qword()),
-      data_len      : Number(st.qword()),
+      cmd : st.dword(),
+      len : st.dword(),
+      image_id : Number(st.qword()),
+      data_offset : Number(st.qword()),
+      data_len : Number(st.qword()),
     }
   }
 
@@ -105,10 +105,10 @@ export class CommandHandler {
       throw new Error("DataError");
     let st = new structHelper_io(data)
     return {
-        cmd         : st.dword(),
-        len         : st.dword(),
-        client_cmd  : st.dword(),
-        data_len    : st.dword(),
+        cmd : st.dword(),
+        len : st.dword(),
+        client_cmd : st.dword(),
+        data_len : st.dword(),
     }
   }
 }
