+++ src/app/Flash.jsx
@@ -1,10 +1,9 @@
 'use client'
-import { useCallback, useState } from 'react'
+import React, { useCallback, useState } from 'react'
 import { CopyToClipboard } from 'react-copy-to-clipboard'
 import Image from 'next/image'
 
-//import { Step, Error, useFastboot } from '@/utils/fastboot'
-import { Step, Error, useQdl } from '@/QDL/qdl'
+import { Step, Error, useQdl } from '@/utils/flash'
 
 import bolt from '@/assets/bolt.svg'
 import cable from '@/assets/cable.svg'
@@ -119,6 +118,13 @@ const errors = {
   },
 }
 
+const detachScripts = [
+  "bus=$(lsusb | grep 05c6:9008 | awk '{print $2}' | sed 's/Bus //;s/^0*//')",
+  "port=$(lsusb -t | grep Driver=qcserial | awk -F'Port ' '{print $2}' | cut -d ':' -f 1)",
+  "echo -n \"$bus-$port:1.0\" | sudo tee /sys/bus/usb/drivers/qcserial/unbind > /dev/null"
+];
+
+const isLinux = navigator.userAgent.toLowerCase().includes('linux');
 
 function LinearProgress({ value, barColor }) {
   if (value === -1 || value > 100) value = 100
@@ -229,13 +235,13 @@ export default function Flash() {
   }
 
   const [copied, setCopied] = useState(false);
-
   const handleCopy = () => {
     setCopied(true);
     setTimeout(() => {
       setCopied(false);
     }, 1000);
   };
+
 
   return (
     <div id="flash" className="relative flex flex-col gap-8 justify-center items-center h-full">
@@ -257,20 +263,33 @@ export default function Flash() {
       </div>
       <span className={`text-3xl dark:text-white font-mono font-light`}>{title}</span>
       <span className={`text-xl dark:text-white px-8 max-w-xl`}>{description}</span>
-      {title === "Lost connection" && (
+      {(title === "Lost connection" || title === "Ready") && isLinux && (
         <>
-          <span className={`text-xl dark:text-white px-8 max-w-xl`}>
-            If you are on Linux, make sure to run the script below in your terminal after plugging in your device.
+          <span className={`text-l dark:text-white px-2`}>
+            It seems that you're on Linux, make sure to run the script below in your terminal after plugging in your device.
           </span>
-          <div className="flex items-center">
-            <pre className="bg-gray-200 dark:bg-gray-800 rounded-md p-4 overflow-x-auto inline-block">
-              <code className="font-mono text-base text-gray-800 dark:text-gray-200 bg-gray-300 dark:bg-gray-700 rounded-md p-3">
-                curl -o- https://bongbui321.github.io/flash/static/detach.sh | bash
-                <CopyToClipboard text="curl -o- https://bongbui321.github.io/flash/static/detach.sh | bash">
-                  <button onClick={handleCopy} className={`bg-${copied ? 'green' : 'blue'}-500 text-white px-1 py-1 rounded-md ml-2 text-sm`}>Copy</button>
-                </CopyToClipboard>
-              </code>
-            </pre>
+          <div className="relative mt-2">
+            <div className="bg-gray-200 dark:bg-gray-800 rounded-md overflow-x-auto">
+              <div className="relative">
+                <pre className="font-mono text-base text-gray-800 dark:text-gray-200 bg-gray-300 dark:bg-gray-700 rounded-md p-5">
+                  {detachScripts.map((line, index) => (
+                    <span key={index} className="block">
+                      {line}
+                    </span>
+                  ))}
+                </pre>
+                <div className="absolute top-2 right-2">
+                  <CopyToClipboard text={detachScripts.join('\n')}>
+                    <button
+                      onClick={handleCopy}
+                      className={`bg-${copied ? 'green' : 'blue'}-500 text-white px-1 py-1 rounded-md ml-2 text-sm`}
+                    >
+                      Copy
+                    </button>
+                  </CopyToClipboard>
+                </div>
+              </div>
+            </div>
           </div>
         </>
       )}
