+++ src/utils/manifest.js
